Preparing branch for merge with master. (cherry picked from commit a729b9c678be527c1f87534ad71d5347c1f5f327)
git-svn-id: https://llvm.org/svn/llvm-project/vmkit/trunk@180527 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/j3/VMCore/JavaClass.h
@@ -468,7 +468,7 @@ public:
 
   /// nbVirtualMethods - The number of virtual methods.
   ///
-  uint16 nbVirtualMethods;
+  uint32 nbVirtualMethods;
   
   /// staticMethods - List of all the static methods defined by this class.
   ///
